added password verification

+++ database.cpp
@@ -1,21 +1,20 @@
 #include "database.h"
 #include <sstream>
 #include <iostream>
-
-
+#include <algorithm>
+#include <openssl/sha.h>
+#include "logging.h"
 
 
 database::database(char* filename) {
     db = NULL;
     open(filename);
     vector<vector<string> > tableExist = query("SELECT name FROM sqlite_master WHERE type='table' AND name='USER';");
-    try {
-        if (tableExist.at(0).at(0).compare("USER")){
-            query("CREATE TABLE USER (name text NOT NULL, location text NOT NULL, age int NOT NULL)");
-        }
-    }
-    catch (const exception &ex){
+    if (std::find(tableExist.at(0).begin(), tableExist.at(0).end(), "USER") == tableExist.at(0).end()){
         query("CREATE TABLE USER (name text NOT NULL, location text NOT NULL, age int NOT NULL)");
+    }
+    if (std::find(tableExist.at(0).begin(), tableExist.at(0).end(), "AUTH") == tableExist.at(0).end()){
+        query("CREATE TABLE AUTH (name text NOT NULL, password int NOT NULL)");
     }
 }
 
@@ -71,32 +70,67 @@ int database::insertUser(user *user) {
           << user->getAge() << ");";
 
     query(sqlSS.str().c_str());
+    setPassword(user);
     
     return 0;
 
 }
 
-user *database::getUser(char* name, user *user) {
-    ostringstream sqlSS;
-    sqlSS << "SELECT * FROM USER WHERE name=\""
-          << name << "\";";
-    vector<vector<string> > buf = query(sqlSS.str().c_str());
-    user->setName(buf[0][0]);
-    user->setLocation(buf[0][1]);
-    user->setAge(atoi(buf[0][2].c_str()));
+user *database::getUser(user requestUser, user *ResponseUser) {
+    if (isAuth(requestUser)) {
+        ostringstream sqlSS;
+        sqlSS << "SELECT * FROM USER WHERE name=\""
+              << requestUser.getName() << "\";";
+        vector<vector<string> > buf = query(sqlSS.str().c_str());
+        ResponseUser->setName(buf[0][0]);
+        ResponseUser->setLocation(buf[0][1]);
+        ResponseUser->setAge(atoi(buf[0][2].c_str()));
+    }
+    else {
+        log.log("not authorised");
+    }
 
-    return (user);
+    return (ResponseUser);
 }
 
-int database::deleteUser(char* name) {
-    ostringstream sqlSS;
-    sqlSS << "DELETE FROM USER WHERE name=\""
-          << name << "\";";
-    query(sqlSS.str().c_str());
+int database::deleteUser(user user) {
+    if (isAuth(user)) {
+        ostringstream sqlSS;
+        sqlSS << "DELETE FROM USER WHERE name=\""
+              << user.getName() << "\";";
+        query(sqlSS.str().c_str());
+    }
+    else {
+        log.log("not authorised");
+    }
     return 0;
 }
 
 void database::close() {
     sqlite3_close(db);
 }
+bool database::setPassword(user *user) {
+    ostringstream sqlSS;
+    sqlSS << "INSERT INTO AUTH (name, password) VALUES (\"" << user->getName() << "\"," << user->getPassword() << ");";
+    query(sqlSS.str().c_str());
+    return true;
+}
+
+bool database::isAuth(user user) {
+    ostringstream sqlSS;
+    unsigned char auth[256];
+    SHA_CTX context;
+    SHA1_Init(&context);
+    char temp[50];
+    sprintf(temp, "%i", user.getPassword());
+    SHA1_Update(&context, temp, 4);
+    SHA1_Final(auth, &context);
+    //cout << "sha1: \"" << auth << "\"" << endl;
+    sqlSS << "SELECT * FROM AUTH WHERE name=\""
+          << user.getName() << "\";";
+    vector<vector<string> > buf = query(sqlSS.str().c_str());
+
+    return (user.getPassword() == atoi(buf[0][1].c_str()));
+}
+
